requirements.txt
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
schedule==1.2.0
dash==2.13.0
plotly==5.18.0